requirements.txt
fastapi
uvicorn
insightface
onnxruntime
numpy
opencv-python-headless
python-multipart
flask
pydantic_core
scikit-image
httpx
boto3
python-dotenv